fix: auth flow bug
