+++ longhorn-studies/app/(app)/(root)/_layout.tsx
@@ -16,7 +16,7 @@ export default function AppLayout() {
   if (!session) {
     // On web, static rendering will stop here as the user is not authenticated
     // in the headless Node process that the pages are rendered in.
-    return <Redirect href="/(app)" />;
+    return <Redirect href="/onboarding" />;
   }
 
   // This layout can be deferred because it's not the root layout.
