+++ longhorn-studies/app/(app)/_layout.tsx
@@ -5,7 +5,7 @@ import { Stack } from 'expo-router';
 export default function Layout() {
   return (
     <Stack>
-      <Stack.Screen name="index" options={{ title: 'Onboarding', headerShown: false }} />
+      <Stack.Screen name="onboarding" options={{ title: 'Onboarding', headerShown: false }} />
       <Stack.Screen name="login" options={{ title: 'Login', presentation: 'modal' }} />
     </Stack>
   );
