+++ longhorn-studies/store/AuthProvider.tsx
@@ -62,16 +62,10 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
   // Sign in with email and password
   const signIn = async (email: string, password: string) => {
     try {
-      const {
-        data: { session, user },
-        error,
-      } = await supabase.auth.signInWithPassword({
+      const { error } = await supabase.auth.signInWithPassword({
         email,
         password,
       });
-
-      setSession(session);
-      setUser(user);
 
       if (error) {
         throw error;
@@ -88,7 +82,12 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
         email,
         password,
       });
-      return { data, error };
+
+      if (error) {
+        throw error;
+      }
+
+      return { data };
     } catch (error) {
       throw error;
     }
@@ -103,6 +102,11 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
   const resetPassword = async (email: string) => {
     try {
       const { error } = await supabase.auth.resetPasswordForEmail(email);
+
+      if (error) {
+        throw error;
+      }
+
       return { error };
     } catch (error) {
       return { error: error as Error };
